refactor(utils): centralize parsePositiveInt to address Gittensory feedback
Move parsePositiveInt from local implementations in src/api/routes.ts, src/github/webhook.ts, and src/orb/webhook.ts to a shared utility in src/utils/json.ts. Update all call sites to import from the shared location and update the test to exercise the production function directly.

Addresses Gittensory review feedback about test coverage and code duplication.
