+++ src/utils/json.ts
@@ -36,3 +36,11 @@ export function repoParts(fullName: string): { owner: string; name: string } {
     name: rest.join("/"),
   };
 }
+
+export function parsePositiveInt(value: string | null | undefined): number | null {
+  if (!value) return null;
+  if (!/^\d+$/.test(value)) return null;
+  const parsed = Number.parseInt(value, 10);
+  if (!Number.isFinite(parsed) || parsed <= 0) return null;
+  return parsed;
+}
